KX-2258 - Code review
